requirements.txt
asgiref==3.3.1
boto3==1.16.28
botocore==1.19.28
coverage==5.3
dj-database-url==0.5.0
Django==3.1.3
django-better-admin-arrayfield==1.4.0
django-compressor==2.4
django-modeltranslation==0.16.1
django-sass-processor==0.8.2
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
libsass==0.20.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
rcssmin==1.0.6
rjsmin==1.1.0
s3transfer==0.3.3
sqlparse==0.4.1
